Move cucumber runtime options management from CucumberSerenityRunner to CucumberWithSerenityRuntime
